Fix 7 WIDE/VM bugs and add 18 targeted tests
VM fixes (vm.cpp):
- Add 22 missing WIDE opcodes: NEW_ARRAY, NEW_MAP, GET_INDEX, SET_INDEX,
  GET_FIELD, SET_FIELD, NEW_CLASS, NEW_INSTANCE, GET_METHOD, NEW_RANGE,
  ITER_INIT, ITER_NEXT, ITER_DONE, PRINT, FIBER_YIELD, FIBER_RESUME,
  TAIL_CALL, AWAIT, THROW, TRY_BEGIN, TRY_END, CLOSURE
- Fix WIDE GET_UPVALUE/SET_UPVALUE: remove wrong nil/closed checks,
  match normal handlers using direct *uv->location
- Fix WIDE RETURN: move fiber done check after frame_count_-- to match
  normal handler order and prevent frame underflow
- Fix WIDE JMP offset: add INST_SIZE compensation for patch_jump's
  WIDE adjustment (was off by 4 bytes)
- Fix TRY_BEGIN/TRY_END: remove bx*4 multiplication, compiler stores
  byte offsets not instruction counts

Compiler fixes (codegen.cpp):
- Remove static_cast<uint8_t> on field/method constant indices passed
  to emit() - truncated indices > 255 causing "Invalid field name"

Tests (test_vm.cpp):
- Add 18 WIDE-specific tests covering all fixed opcodes with >255
  register context: arithmetic, comparisons, jumps, closures, classes,
  arrays, maps, iterators, fibers, try/catch

+++ src/compiler/codegen.cpp
@@ -365,7 +365,7 @@ void CodeGenerator::compile_call(CallExpr* node, int reg) {
         uint16_t method_const = make_identifier_constant(field->field);
         // Allocate callee register
         int callee = alloc_register();
-        emit(op_byte(Opcode::GET_FIELD), callee, object_reg, static_cast<uint8_t>(method_const));
+        emit(op_byte(Opcode::GET_FIELD), callee, object_reg, method_const);
 
         // Compile args
         std::vector<int> arg_regs;
@@ -426,14 +426,14 @@ void CodeGenerator::compile_index(IndexExpr* node, int reg) {
 void CodeGenerator::compile_field_access(FieldAccessExpr* node, int reg) {
     int obj = compile_expr(node->object);
     uint16_t field_const = make_identifier_constant(node->field);
-    emit(op_byte(Opcode::GET_FIELD), reg, obj, static_cast<uint8_t>(field_const));
+    emit(op_byte(Opcode::GET_FIELD), reg, obj, field_const);
 }
 
 void CodeGenerator::compile_field_set(FieldSetExpr* node, int reg) {
     int obj = compile_expr(node->object);
     int val = compile_expr(node->value);
     uint16_t field_const = make_identifier_constant(node->field);
-    emit(op_byte(Opcode::SET_FIELD), obj, static_cast<uint8_t>(field_const), val);
+    emit(op_byte(Opcode::SET_FIELD), obj, field_const, val);
     emit(op_byte(Opcode::MOVE), reg, val, 0);
 }
 
@@ -779,7 +779,7 @@ void CodeGenerator::compile_class(ClassStmt* node) {
         int method_reg = alloc_register();
         emit_bx(op_byte(Opcode::CLOSURE), method_reg, func_const);
         uint16_t method_const = make_identifier_constant(method->name);
-        emit(op_byte(Opcode::SET_FIELD), class_reg, static_cast<uint8_t>(method_const), method_reg);
+        emit(op_byte(Opcode::SET_FIELD), class_reg, method_const, method_reg);
         free_register();
     }
